refactor count imports
