+++ src/counts/index.ts
@@ -1,0 +1,4 @@
+import { embezzlerCount } from "./embezzler";
+import { digitizedMonstersRemaining } from "./digitize";
+
+export { digitizedMonstersRemaining, embezzlerCount };
